test: Add a test for site plugins with two plugins
This test is a small variation on the example in the manual but with two
plugins. One of those plugins comes with a test that really depends on
this plugin (so it is expected to succeed whether the other plugin is
available or not). It shows that a test will fail if it launches the
application before all the plugins have been fully built, even if they
are not required.

Signed-off-by: Samuel Hym <samuel.hym@rustyne.lautre.net>